requirements.txt
biopython==1.76
pandas==1.0.1
Click==7.0
